- **Enable easy profiling**
      python -m cProfile yanki/cli.py build -o lp.apkg asl/*.deck
  

- **Speed: cache `raw_to_html()`**
  This change takes the overall time when profiling down from 49 seconds
  to 9 seconds.
  
  In actual run time:
  
      ❯ hyperfine -w 1 -n main -p 'git switch main' 'yanki build -o lp.apkg asl/*.deck' -n profile -p 'git switch profile' 'yanki build -o lp.apkg asl/*.deck'
      Benchmark 1: main
        Time (mean ± σ):     22.743 s ±  1.535 s    [User: 20.747 s, System: 1.722 s]
        Range (min … max):   20.852 s … 25.832 s    10 runs
  
        Warning: Statistical outliers were detected. Consider re-running this benchmark on a quiet PC without any interferences from other programs. It might help to use the '--warmup' or '--prepare' options.
  
      Benchmark 2: profile
        Time (mean ± σ):      3.188 s ±  0.061 s    [User: 2.650 s, System: 0.517 s]
        Range (min … max):    3.116 s …  3.321 s    10 runs
  
      Summary
        'profile' ran
          7.13 ± 0.50 times faster than 'main'